feat: record authorization

+++ app/(feature)/record/api/[id]/route.ts
@@ -3,6 +3,7 @@ import { Collection } from '../../_db'
 import { formatFormRequest } from "../recordFormatter"
 import { revalidateTag } from "next/cache"
 import { RecordFormRequestSchema } from "../../_db/schema/RecordFormRequestSchema"
+import { getUserSession } from "@/auth/_lib/getUserSession"
 
 export interface Params {
   params: {
@@ -12,6 +13,10 @@ export interface Params {
 
 export async function DELETE(req: Request, { params: { id } }: Params) {
   try {
+    const session = await getUserSession()
+    if (!session?.user) {
+      return errorResponse({status: 401})
+    }
     await Collection.deleteData(id)
     revalidateTag('records')
     return successResponse()
@@ -24,9 +29,16 @@ export async function DELETE(req: Request, { params: { id } }: Params) {
 export async function PUT(req: Request, params: Params) { 
   const { params: { id } } = params;
   try {
+    const session = await getUserSession()
+    if (!session?.user) {
+      return errorResponse({status: 401})
+    }
     const json = await req.json()
-    const parsedData = RecordFormRequestSchema.parse(json)
-    const data = await formatFormRequest(parsedData)
+    const parseDataWithUser = {
+      ...RecordFormRequestSchema.parse(json),
+      user: session.user._id
+    }
+    const data = await formatFormRequest(parseDataWithUser)
     await Collection.updateData(id, data)
     revalidateTag('records')
     return successResponse({data})
